Fix for Blood Moon and ETB replacement effects
As per IMA release notes for Magus of the Moon:
If a nonbasic land has an ability that applies "as [this land] enters the battlefield"
or that causes it to enter the battlefield tapped or with counters,
the land will lose that ability before it applies.
This is a change from previous rules.